FIX: interest_and_principal_payments formula and tests fixed
There was a mismatch betwenn the `fv` and the cummulative sum of the principal
payments at the end of the amortization period because of the rounding
errors. The fix was to make the final payment as the sum of the
remaining value in the `pv` after the cummulative sum of the principal
payments and the `fv`.

There may be inconsistencies in the values. So I am in the lookout for
errors in my implementation.

Changed the test to match the new results. Used MacOS Numbers formulas
to get the results. will be creating more cases.