FEAT: Add SaleYearBalloonPayment() calculation.
returns the calculated balloon payment at the time of sale of the
property. In the investment_analysis module, there is a need to know the
balloon payment at the time of sale of the property, in case the sale is
not done at the end of the term but sooner.

Modified the tests to make sure that new method names are taken.